sites api views refactoring: move sass into v2 apis + tasks helpers
 - `import_course_on_site_creation_after_transaction` prep for future use
