+++ openedx/core/djangoapps/appsembler/sites/tasks.py
@@ -1,6 +1,7 @@
 """
 This file contains celery tasks for contentstore views
 """
+import beeline
 import logging
 import requests
 import os
@@ -12,6 +13,7 @@ from tempfile import NamedTemporaryFile, TemporaryDirectory
 from celery.task import task
 from celery.utils.log import get_task_logger
 
+from django.db import transaction
 from django.conf import settings
 
 from student.models import CourseEnrollment
@@ -28,6 +30,28 @@ from xmodule.modulestore.exceptions import ItemNotFoundError
 from xmodule.modulestore.xml_importer import import_course_from_xml
 
 LOGGER = get_task_logger(__name__)
+
+
+def import_course_on_site_creation_after_transaction(organization):
+    """
+    Clone the course after Database transaction is committed, only if the import feature is enabled.
+
+    :param organization: Organization to create the course for.
+    :return bool: Whether the course is scheduled for creation or not.
+    """
+    if settings.FEATURES.get("APPSEMBLER_IMPORT_DEFAULT_COURSE_ON_SITE_CREATION", False):
+        beeline.add_context_field("default_course_on_site_creation_flag", True)
+
+        def import_task_on_commit():
+            """
+            Run the import task after the commit to avoid Organization.DoesNotExist error on the Celery.
+            """
+            import_course_on_site_creation_apply_async(organization)
+
+        transaction.on_commit(import_task_on_commit)
+        return True
+
+    return False
 
 
 def import_course_on_site_creation_apply_async(organization):
