+++ openedx/core/djangoapps/appsembler/sites/urls.py
@@ -2,8 +2,7 @@ from django.conf import settings
 from django.conf.urls import url, include
 from rest_framework.routers import DefaultRouter
 
-from openedx.core.djangoapps.appsembler.sites.api import (
-    CompileSassView,
+from .api import (
     CustomDomainView,
     DomainAvailabilityView,
     DomainSwitchView,
@@ -15,6 +14,8 @@ from openedx.core.djangoapps.appsembler.sites.api import (
     SiteViewSet,
     UsernameAvailabilityView,
 )
+
+from . import api_v2
 
 # Create a router and register our viewsets with it.
 router = DefaultRouter()
@@ -30,8 +31,8 @@ urlpatterns = [
     url(r'^domain/(?P<subdomain>[\w.@+-]+)/', DomainAvailabilityView.as_view()),
     url(r'^custom_domain/', CustomDomainView.as_view()),
     url(r'^domain_switch/', DomainSwitchView.as_view()),
-    url(r'^compile_sass/', CompileSassView.as_view(), name='compile_sass'),
     url(r'^register/', SiteCreateView.as_view(), name='tahoe_site_creation'),
+    url(r'^v2/compile_sass/', api_v2.CompileSassView.as_view(), name='tahoe_compile_sass'),
     url(r'^', include(router.urls)),
 ]
